+++ frontend/src/app/address-book/page.tsx
@@ -1,15 +1,15 @@
 "use client";
 import { useEffect, useState } from "react";
-import useTheme from "../components/ui_components/hooks/useTheme";
 import logoImage from "../../../public/assets/logo.svg";
 import { AddressBookResource } from "../types";
 import { addressBookResources, searchResources } from "../utils";
 import { useDebounce } from "../hooks";
 import Image from "next/image";
 import AddressTable from "./address-table";
-import ThemeSwitch from "../components/header/ThemeSwitch";
+import ThemeSwitch from "../components/ui_components/header/ThemeSwitch";
 import Search from "svg/Search";
 import Link from "next/link";
+import useTheme from "../hooks/useTheme";
 
 export default function Page() {
   const { theme, changeTheme } = useTheme();
@@ -51,12 +51,7 @@ export default function Page() {
                 addressbook
               </h4>
             </div>
-            <ThemeSwitch
-              dimension="w-[3rem] h-[3rem] md:w-[3.5rem] md:h-[3.5rem] lg:w-[4rem] lg:h-[4rem]"
-              className="grid"
-              action={changeTheme}
-              theme={theme}
-            />
+            <ThemeSwitch className="grid" action={changeTheme} theme={theme} />
           </div>
           <div className="relative mx-auto text-[--headings] md:w-fit">
             <input
