+++ frontend/src/app/components/ui_components/AccountBalance.tsx
@@ -1,9 +1,8 @@
 import { useContractRead } from "@starknet-react/core";
-import { useEffect, useState } from "react";
-import { ETH_SEPOLIA, STRK_SEPOLIA } from "./utils/constant";
 import Erc20Abi from "../../abi/token.abi.json";
-import { formatCurrency } from "./utils/currency";
 import RightChevron from "svg/RightChevron";
+import { ETH_SEPOLIA, STRK_SEPOLIA } from "@/app/utils/constant";
+import { formatCurrency } from "@/app/utils/currency";
 
 type Props = {
   address: string;
