+++ frontend/src/app/components/ui_components/header/AddTokenModal.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useState } from "react";
 import { useConnect } from "@starknet-react/core";
-import GenericModal from "../ui_components/GenericModal";
+import GenericModal from "../GenericModal";
 import Close from "svg/Close";
 
 const AddTokenModal = () => {
