+++ frontend/src/app/utils/currency.ts
